+++ spa/src/app/files/hca-table-files/hca-table-files.component.html
@@ -1,5 +1,5 @@
 <div class="wrapper">
-    <div class="hca-table">
+    <div class="hca-table files">
         <mat-table *ngIf="pagination$ | async; let pagination" #table [dataSource]="tableElementDataSource" matSort
                    (matSortChange)="sortTable(pagination, $event)">
             <ng-container matColumnDef="fileName">
@@ -16,7 +16,8 @@
                          matTooltipPosition="below"
                          matTooltipShowDelay="{{tooltipShowDelay}}"
                          matTooltipClass="{{getTooltipClass(element.fileName)}}"
-                         matTooltipDisabled="{{isDisabled(file)}}">{{element.fileName}}
+                         matTooltipDisabled="{{isDisabled(file)}}"
+                         [ngClass]="getTruncatedClass(element.fileName)">{{element.fileName}}
                     </div>
                 </mat-cell>
             </ng-container>
@@ -84,7 +85,8 @@
                          matTooltipPosition="below"
                          matTooltipShowDelay="{{tooltipShowDelay}}"
                          matTooltipClass="hca-tooltip"
-                         matTooltipDisabled="{{isDisabled(specimen)}}">{{element.specimenId}}
+                         matTooltipDisabled="{{isDisabled(specimen)}}"
+                         [ngClass]="getTruncatedClass(element.fileName)">{{element.specimenId}}
                     </div>
                 </mat-cell>
             </ng-container>
@@ -228,16 +230,6 @@
             <mat-header-row *matHeaderRowDef="displayedColumns" [ngClass]="getHeaderClass()"></mat-header-row>
             <mat-row *matRowDef="let row; columns: displayedColumns; let i = index" [ngClass]="getRowClass(i)"></mat-row>
         </mat-table>
-        <div *ngIf="pagination$ | async; let pagination" class="pagination">
-            <ng-container *ngIf="pagination.current_page">
-                <button mat-icon-button (click)="previousPageSelected(pagination)" [disabled]="pagination.current_page === 1">
-                    <mat-icon>chevron_left</mat-icon>
-                </button>
-                <span class="fontsize-xs">Page {{pagination.current_page}} of {{getPageCount(pagination)}}</span>
-                <button mat-icon-button (click)="nextPageSelected(pagination)" [disabled]="pagination.current_page === getPageCount(pagination)">
-                    <mat-icon>chevron_right</mat-icon>
-                </button>
-            </ng-container>
-        </div>
+        <hca-table-pagination></hca-table-pagination>
     </div>
 </div>
