+++ spa/src/app/files/hca-table-files/hca-table-files.component.ts
@@ -17,10 +17,7 @@ import { Subject } from "rxjs/Subject";
 // App dependencies
 import { AppState } from "../../_ngrx/app.state";
 import { selectPagination, selectTableData } from "../_ngrx/file.selectors";
-import {
-    FetchPagedOrSortedTableDataRequestAction, TableNextPageAction,
-    TablePreviousPageAction
-} from "../_ngrx/table/table.actions";
+import { FetchPagedOrSortedTableDataRequestAction } from "../_ngrx/table/table.actions";
 import { PaginationModel } from "../table/pagination.model";
 import { TableParamsModel } from "../table/table-params.model";
 
@@ -129,6 +126,18 @@ export class HCATableFilesComponent implements OnInit, AfterViewInit {
     }
 
     /**
+     * Returns class truncate if table data is not spaced.
+     * @param {string} tableData
+     * @returns {string}
+     */
+    public getTruncatedClass(tableData: string): string {
+
+        if ( tableData.indexOf(" ") == -1 ) {
+            return "truncate";
+        }
+    }
+
+    /**
      * Returns false if the text is longer than its container.
      * If false, an ellipsis has been applied to the text.
      * @param el
@@ -137,51 +146,6 @@ export class HCATableFilesComponent implements OnInit, AfterViewInit {
     public isDisabled(el) {
 
         return !( el.scrollWidth > el.clientWidth );
-    }
-
-    /**
-     * Called when table next page selected.
-     *
-     * @param {PaginationModel} pm
-     */
-    public nextPageSelected(pm: PaginationModel) {
-
-        if ( !this.hasNext(pm) ) {
-            return;
-        }
-
-        let tableParamsModel = {
-
-            search_after: pm.search_after,
-            search_after_uid: pm.search_after_uid,
-            size: pm.size,
-            sort: pm.sort,
-            order: pm.order
-        };
-
-        this.store.dispatch(new TableNextPageAction(tableParamsModel));
-    }
-
-    /**
-     * Called when table previous page selected.
-     */
-    public previousPageSelected(pm: PaginationModel) {
-
-
-        if ( !this.hasPrevious(pm) ) {
-            return;
-        }
-
-        let tableParamsModel = {
-
-            search_before: pm.search_before,
-            search_before_uid: pm.search_after_uid,
-            size: pm.size,
-            sort: pm.sort,
-            order: pm.order
-        };
-
-        this.store.dispatch(new TablePreviousPageAction(tableParamsModel));
     }
 
     /**
@@ -196,67 +160,6 @@ export class HCATableFilesComponent implements OnInit, AfterViewInit {
             size: pm.size,
             sort: sort.active,
             order: sort.direction
-        };
-
-        this.store.dispatch(new FetchPagedOrSortedTableDataRequestAction(tableParamsModel));
-    }
-
-    /**
-     * Check if there is a next page.
-     *
-     * @param {PaginationModel} pm
-     * @returns {boolean}
-     */
-    public hasNext(pm: PaginationModel): boolean {
-        // return (pm.from + pm.count) < pm.total;
-        return pm.search_after !== null;
-    }
-
-    /**
-     * Check if there is a previous page.
-     *
-     * @param {PaginationModel} pm
-     * @returns {boolean}
-     */
-    public hasPrevious(pm: PaginationModel): boolean {
-        // return (pm.from > 1);
-        return pm.search_before !== null;
-    }
-
-    /**
-     * Return the total number of pages.
-     *
-     * @param {PaginationModel} pm
-     * @returns {number}
-     */
-    getPageCount(pm: PaginationModel) {
-        return Math.ceil(pm.total / pm.size);
-    }
-
-    getPages(pm: PaginationModel): number[] {
-
-        let pages = [];
-        let pageCount = this.getPageCount(pm);
-
-        for ( let i = 1; i <= pageCount; i++ ) {
-            pages.push(i);
-        }
-
-        return pages;
-    }
-
-    /**
-     * Sets the number of rows per page.
-     *
-     * @param {PaginationModel} pm
-     * @returns {number} pageSize
-     */
-    public setPageSize(pm: PaginationModel, pageSize: number) {
-
-        let tableParamsModel: TableParamsModel = {
-            size: pageSize,
-            sort: pm.sort,
-            order: pm.order
         };
 
         this.store.dispatch(new FetchPagedOrSortedTableDataRequestAction(tableParamsModel));
