+++ spa/src/app/files/hca-table-projects/hca-table-projects.component.html
@@ -1,5 +1,5 @@
 <div class="wrapper">
-    <div class="hca-table">
+    <div class="hca-table projects">
         <mat-table *ngIf="pagination$ | async; let pagination" #table [dataSource]="tableElementDataSource" matSort
                    (matSortChange)="sortTable(pagination, $event)">
             <ng-container matColumnDef="projectTitle">
@@ -175,18 +175,7 @@
             <mat-row *matRowDef="let row; columns: displayedColumns; let i = index"
                      [ngClass]="getRowClass(i)"></mat-row>
         </mat-table>
-        <div *ngIf="pagination$ | async; let pagination" class="pagination">
-            <ng-container *ngIf="pagination.current_page">
-                <button mat-icon-button (click)="previousPageSelected(pagination)"
-                        [disabled]="pagination.current_page === 1">
-                    <mat-icon>chevron_left</mat-icon>
-                </button>
-                <span class="fontsize-xs">Page {{pagination.current_page}} of {{getPageCount(pagination)}}</span>
-                <button mat-icon-button (click)="nextPageSelected(pagination)" [disabled]="pagination.current_page === getPageCount(pagination)">
-                    <mat-icon>chevron_right</mat-icon>
-                </button>
-            </ng-container>
-        </div>
+        <hca-table-pagination></hca-table-pagination>
     </div>
 </div>
