+++ spa/src/app/files/hca-table-projects/hca-table-projects.component.ts
@@ -6,7 +6,6 @@
  */
 
 // Core dependencies
-import { trigger, state, style, animate, transition } from "@angular/animations";
 import { AfterViewInit, ChangeDetectorRef, Component, ElementRef, Inject, Input, OnInit } from "@angular/core";
 import { DataSource } from "@angular/cdk/collections";
 import { Sort } from "@angular/material";
@@ -21,10 +20,7 @@ import { AppState } from "../../_ngrx/app.state";
 import { FileFacetSelectedEvent } from "../file-facets/file-facet.events";
 import { selectPagination, selectTableData } from "../_ngrx/file.selectors";
 import { SelectFileFacetAction } from "../_ngrx/file-facet-list/file-facet-list.actions";
-import {
-    FetchPagedOrSortedTableDataRequestAction, TableNextPageAction,
-    TablePreviousPageAction
-} from "../_ngrx/table/table.actions";
+import { FetchPagedOrSortedTableDataRequestAction } from "../_ngrx/table/table.actions";
 import { FileFacet } from "../shared/file-facet.model";
 import { PaginationModel } from "../table/pagination.model";
 import { TableParamsModel } from "../table/table-params.model";
@@ -131,51 +127,6 @@ export class HCATableProjectsComponent implements OnInit, AfterViewInit {
     }
 
     /**
-     * Called when table next page selected.
-     *
-     * @param {PaginationModel} pm
-     */
-    public nextPageSelected(pm: PaginationModel) {
-
-        if ( !this.hasNext(pm) ) {
-            return;
-        }
-
-        let tableParamsModel = {
-
-            search_after: pm.search_after,
-            search_after_uid: pm.search_after_uid,
-            size: pm.size,
-            sort: pm.sort,
-            order: pm.order
-        };
-
-        this.store.dispatch(new TableNextPageAction(tableParamsModel));
-    }
-
-    /**
-     * Called when table previous page selected.
-     */
-    public previousPageSelected(pm: PaginationModel) {
-
-
-        if ( !this.hasPrevious(pm) ) {
-            return;
-        }
-
-        let tableParamsModel = {
-
-            search_before: pm.search_before,
-            search_before_uid: pm.search_after_uid,
-            size: pm.size,
-            sort: pm.sort,
-            order: pm.order
-        };
-
-        this.store.dispatch(new TablePreviousPageAction(tableParamsModel));
-    }
-
-    /**
      * Sort the table given the sort param and the order.
      *
      * @param {PaginationModel} pm
@@ -187,67 +138,6 @@ export class HCATableProjectsComponent implements OnInit, AfterViewInit {
             size: pm.size,
             sort: sort.active,
             order: sort.direction
-        };
-
-        this.store.dispatch(new FetchPagedOrSortedTableDataRequestAction(tableParamsModel));
-    }
-
-    /**
-     * Check if there is a next page.
-     *
-     * @param {PaginationModel} pm
-     * @returns {boolean}
-     */
-    public hasNext(pm: PaginationModel): boolean {
-        // return (pm.from + pm.count) < pm.total;
-        return pm.search_after !== null;
-    }
-
-    /**
-     * Check if there is a previous page.
-     *
-     * @param {PaginationModel} pm
-     * @returns {boolean}
-     */
-    public hasPrevious(pm: PaginationModel): boolean {
-        // return (pm.from > 1);
-        return pm.search_before !== null;
-    }
-
-    /**
-     * Return the total number of pages.
-     *
-     * @param {PaginationModel} pm
-     * @returns {number}
-     */
-    getPageCount(pm: PaginationModel) {
-        return Math.ceil(pm.total / pm.size);
-    }
-
-    getPages(pm: PaginationModel): number[] {
-
-        let pages = [];
-        let pageCount = this.getPageCount(pm);
-
-        for ( let i = 1; i <= pageCount; i++ ) {
-            pages.push(i);
-        }
-
-        return pages;
-    }
-
-    /**
-     * Sets the number of rows per page.
-     *
-     * @param {PaginationModel} pm
-     * @returns {number} pageSize
-     */
-    public setPageSize(pm: PaginationModel, pageSize: number) {
-
-        let tableParamsModel: TableParamsModel = {
-            size: pageSize,
-            sort: pm.sort,
-            order: pm.order
         };
 
         this.store.dispatch(new FetchPagedOrSortedTableDataRequestAction(tableParamsModel));
