+++ spa/src/app/files/hca-table/hca-table.component.html
@@ -241,16 +241,6 @@
             <mat-row *matRowDef="let row; columns: displayedColumns; let i = index"
                      [ngClass]="getRowClass(i)"></mat-row>
         </mat-table>
-        <div *ngIf="pagination$ | async; let pagination" class="pagination">
-            <ng-container *ngIf="pagination.current_page">
-                <button mat-icon-button (click)="previousPageSelected(pagination)" [disabled]="pagination.current_page === 1">
-                    <mat-icon>chevron_left</mat-icon>
-                </button>
-                <span class="fontsize-s">Page {{pagination.current_page}} of {{getPageCount(pagination)}}</span>
-                <button mat-icon-button (click)="nextPageSelected(pagination)" [disabled]="pagination.current_page === getPageCount(pagination)">
-                    <mat-icon>chevron_right</mat-icon>
-                </button>
-            </ng-container>
-        </div>
+        <hca-table-pagination></hca-table-pagination>
     </div>
 </div>
